Updated the function gaussElimination()
Earlier, the function worked only for integer valued matrices, but with this update, it is working for real numbers as well.

+++ library.cpp
@@ -155,7 +155,7 @@ int *Matrix::gaussElimination()
 	for(int i=0;i<c;i++)
 		for(int j=i+1;j<r;j++)
 		{
-			int x=m[j][i]/m[i][i];
+			float x=m[j][i]/m[i][i];
 			*ptr=x;
 			ptr++;
 			for(int k=0;k<c;k++)
